First pass of pointer conversion; still asserting
I learned from [this page](https://wiki.unrealengine.com/How_To_Prevent_Crashes_Due_To_Dangling_Actor_Pointers) that I was segfaulting in UnrealMovementComponent::Update() because storing raw pointers to UObjects is treacherous. I converted everything in UnrealMovementComponentManager to use Shared and Weak pointers for Actors and Characters. 

This seemed all fine and dandy, but now I'm running into an infuriating ensure assert where ACharacter can't be constructed because it's already registered. What is perplexing about this is that I can construct the AgentCharacters just fine in GalavantUnrealMain::Tick(), but as soon as I do it in UnrealMovementComponent::Update() with the same exact parameters everything goes to shit.

I'm at a loss as to what is broken as it seems impossible that my changes could've broken things the way they've been broken.